sih_ml/requirements.txt
attrs==19.3.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.3
Django==2.0.13
django-tree==0.3.2
djangorestframework==3.11.0
docx2txt==0.8
en-core-web-sm==2.3.1
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
jsonschema==3.2.0
murmurhash==1.0.2
nltk==3.5
numpy==1.19.0
pandas==0.25.3
pdfminer.six==20200517
plac==1.1.3
preshed==3.0.2
pycryptodome==3.9.8
pyresparser==1.0.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
spacy==2.3.1
sqlparse==0.3.1
srsly==2.2.0
thinc==7.4.1
tqdm==4.47.0
urllib3==1.25.9
wasabi==0.7.0
zipp==3.1.0
